move interpreter to own file
