+++ executor/src/witgen/jit/mod.rs
@@ -3,6 +3,7 @@ mod block_machine_processor;
 mod compiler;
 mod effect;
 pub(crate) mod function_cache;
+mod interpreter;
 mod symbolic_expression;
 mod variable;
 pub(crate) mod witgen_inference;
